Moved _make_example_dict_equal_fn to utils.test_util
