+++ tensorflow_data_validation/coders/csv_decoder_test.py
@@ -28,7 +28,7 @@ from apache_beam.testing import util
 import numpy as np
 from tensorflow_data_validation import types
 from tensorflow_data_validation.coders import csv_decoder
-from tensorflow_data_validation.coders.decoder_test_util import _make_example_dict_equal_fn
+from tensorflow_data_validation.utils import test_util
 from tensorflow_data_validation.types_compat import Callable, List
 
 
@@ -52,7 +52,7 @@ class CSVDecoderTest(absltest.TestCase):
                 csv_decoder.DecodeCSV(column_names=column_names))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_missing_values(self):
     input_lines = ['1,,hello',
@@ -71,7 +71,7 @@ class CSVDecoderTest(absltest.TestCase):
                 csv_decoder.DecodeCSV(column_names=column_names))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_with_int_and_float_in_same_column(self):
     input_lines = ['2,1.5',
@@ -88,7 +88,7 @@ class CSVDecoderTest(absltest.TestCase):
                 csv_decoder.DecodeCSV(column_names=column_names))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_with_int_and_string_in_same_column(self):
     input_lines = ['2,abc',
@@ -105,7 +105,7 @@ class CSVDecoderTest(absltest.TestCase):
                 csv_decoder.DecodeCSV(column_names=column_names))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_with_float_and_string_in_same_column(self):
     input_lines = ['2.3,abc',
@@ -122,7 +122,7 @@ class CSVDecoderTest(absltest.TestCase):
                 csv_decoder.DecodeCSV(column_names=column_names))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_with_unicode(self):
     input_lines = [u'1,שקרכלשהו,22.34,text field']
@@ -140,7 +140,7 @@ class CSVDecoderTest(absltest.TestCase):
                 csv_decoder.DecodeCSV(column_names=column_names))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_csv_record_with_quotes(self):
     input_lines = ['1,"ab,cd,ef"',
@@ -157,7 +157,7 @@ class CSVDecoderTest(absltest.TestCase):
                 csv_decoder.DecodeCSV(column_names=column_names))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_with_space_delimiter(self):
     input_lines = ['1 "ab,cd,ef"',
@@ -175,7 +175,7 @@ class CSVDecoderTest(absltest.TestCase):
                                       delimiter=' '))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_with_tab_delimiter(self):
     input_lines = ['1\t"this is a \ttext"',
@@ -193,7 +193,7 @@ class CSVDecoderTest(absltest.TestCase):
                                       delimiter='\t'))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_negative_values(self):
     input_lines = ['-34', '45']
@@ -207,7 +207,7 @@ class CSVDecoderTest(absltest.TestCase):
                 csv_decoder.DecodeCSV(column_names=column_names))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_int64_max(self):
     input_lines = ['34', str(sys.maxsize)]
@@ -221,7 +221,7 @@ class CSVDecoderTest(absltest.TestCase):
                 csv_decoder.DecodeCSV(column_names=column_names))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_large_int_categorical_pos(self):
     input_lines = ['34', str(sys.maxsize+1)]
@@ -235,7 +235,7 @@ class CSVDecoderTest(absltest.TestCase):
                 csv_decoder.DecodeCSV(column_names=column_names))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_large_int_categorical_neg(self):
     input_lines = ['34', str(-(sys.maxsize+2))]
@@ -249,7 +249,7 @@ class CSVDecoderTest(absltest.TestCase):
                 csv_decoder.DecodeCSV(column_names=column_names))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_large_int_categorical_pos_and_neg(self):
     input_lines = [str(sys.maxsize+1), str(-(sys.maxsize+2))]
@@ -263,7 +263,7 @@ class CSVDecoderTest(absltest.TestCase):
                 csv_decoder.DecodeCSV(column_names=column_names))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_empty_row(self):
     input_lines = [',,',
@@ -282,7 +282,7 @@ class CSVDecoderTest(absltest.TestCase):
                 csv_decoder.DecodeCSV(column_names=column_names))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_skip_blank_line(self):
     input_lines = ['',
@@ -297,7 +297,7 @@ class CSVDecoderTest(absltest.TestCase):
                 csv_decoder.DecodeCSV(column_names=column_names))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_consider_blank_line(self):
     input_lines = ['',
@@ -315,7 +315,7 @@ class CSVDecoderTest(absltest.TestCase):
                                       skip_blank_lines=False))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_skip_blank_line_single_column(self):
     input_lines = ['',
@@ -329,7 +329,7 @@ class CSVDecoderTest(absltest.TestCase):
                 csv_decoder.DecodeCSV(column_names=column_names))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_consider_blank_line_single_column(self):
     input_lines = ['',
@@ -345,7 +345,7 @@ class CSVDecoderTest(absltest.TestCase):
                                       skip_blank_lines=False))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_empty_csv(self):
     input_lines = []
@@ -356,7 +356,7 @@ class CSVDecoderTest(absltest.TestCase):
                 csv_decoder.DecodeCSV(column_names=[]))
       util.assert_that(
           result,
-          _make_example_dict_equal_fn(self, expected_result))
+          test_util._make_example_dict_equal_fn(self, expected_result))
 
   def test_csv_decoder_invalid_row(self):
     input_lines = ['1,2.0,hello',
@@ -370,7 +370,7 @@ class CSVDecoderTest(absltest.TestCase):
                   csv_decoder.DecodeCSV(column_names=column_names))
         util.assert_that(
             result,
-            _make_example_dict_equal_fn(self, None))
+            test_util._make_example_dict_equal_fn(self, None))
 
 
 if __name__ == '__main__':
