--- tensorflow_data_validation/coders/decoder_test_util.py
@@ -1,58 +1,0 @@
-# Copyright 2018 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Utils for decoder tests."""
-
-from __future__ import absolute_import
-from __future__ import division
-
-from __future__ import print_function
-
-from apache_beam.testing import util
-import numpy as np
-
-
-def _make_example_dict_equal_fn(
-    test,
-    expected
-):
-  """Makes a matcher function for comparing the example dict.
-
-  Args:
-    test: test case object.
-    expected: the expected example dict.
-
-  Returns:
-    A matcher function for comparing the example dicts.
-  """
-
-  def _matcher(actual):
-    """Matcher function for comparing the example dicts."""
-    try:
-      # Check number of examples.
-      test.assertEqual(len(actual), len(expected))
-
-      for i in range(len(actual)):
-        for key in actual[i]:
-          # Check each feature value.
-          if isinstance(expected[i][key], np.ndarray):
-            test.assertEqual(actual[i][key].dtype, expected[i][key].dtype)
-            np.testing.assert_equal(actual[i][key], expected[i][key])
-          else:
-            test.assertEqual(actual[i][key], expected[i][key])
-
-    except AssertionError as e:
-      raise util.BeamAssertException('Failed assert: ' + str(e))
-
-  return _matcher
